requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
ollama>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# UI
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
